feat: [US-002] - [Remove the wrapper-only layer and preserve behavioral regression coverage]
