+++ docs/processes/development-guide-relevant-files.md
@@ -76,7 +76,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `ui/.storybook/dashboard-story-runtime.tsx` | Dashboard Storybook runtime seam | Dashboard stories that exercise live selection or timeline behavior should seed the current `features/current-selection` and `features/timeline` stores through this decorator instead of mocking widget-local state, so Storybook browser checks stay aligned with the production dashboard store shape. |
 | `ui/src/features/export/export-factory-dialog.stories.tsx` | Export dialog browser verification seam | Dashboard export-dialog keyboard and validation smoke should live in this Storybook story so browser-backed UI proof stays close to the shared dialog implementation instead of being recreated in ad hoc scripts. |
 | `ui/scripts/verify-import-export-storybook-responsive.mjs` | Storybook responsive dialog verification seam | Breakpoint-sensitive dialog evidence for the import/export lane should stay in this Playwright script so the documented `390x844`, `768x1024`, and `1440x900` checks are executable against built Storybook `iframe.html` stories instead of remaining doc-only claims. |
-| `ui/src/features/workflow-activity/dashboard-import-preview-dialog.stories.tsx` | Import preview browser verification seam | Import-preview keyboard and dialog-surface browser checks should stay in this focused Storybook story so reviewers can rerun the import dialog proof without booting the full graph flow. |
+| `ui/src/features/import/dashboard-import-preview-dialog.stories.tsx` | Import preview browser verification seam | Import-preview keyboard and dialog-surface browser checks should stay in this focused Storybook story beside the dashboard-owned import dialog so reviewers can rerun the import dialog proof without booting the full graph flow. |
 | `ui/src/features/workflow-activity/mutation-dialog.test.tsx` | Shared mutation-dialog verification seam | Import/export dialog cleanup that depends on shared overlay semantics should keep the direct accessible-title and description proof here so future dialog work does not fall back to architecture-doc reasoning alone. |
 | `ui/src/components/ui/dialog.tsx` | Shared shadcn dialog primitive seam | Dashboard mutation overlays that need repository-standard dialog chrome should compose this shared Radix-backed primitive, and flows that must block dismissal during in-flight mutations should use its labeled and disableable close control instead of reintroducing bespoke overlay wrappers. |
 | `ui/src/features/bento/dashboard-bento.tsx` | Dashboard-level overlay composition seam | Dashboard overlays that are triggered from bento widgets but should not be clipped or owned by a widget body should keep their controller state here and render dialog siblings beside `AgentBentoLayout`. |
