+++ ui/src/features/workflow-activity/index.ts
@@ -1,5 +1,4 @@
 export * from "./current-activity-import-controller";
-export * from "./dashboard-import-preview-dialog";
 export * from "./dashboard-flow-axis-legend";
 export * from "./react-flow-current-activity-card";
 export * from "./workflow-activity-bento-card";
